+++ lib/src/main/java/org/asamk/signal/manager/internal/ManagerImpl.java
@@ -1045,7 +1045,10 @@ public class ManagerImpl implements Manager {
         }
         if (!message.attachments().isEmpty()) {
             final var uploadedAttachments = context.getAttachmentHelper()
-                    .uploadAttachments(message.attachments(), message.voiceNote());
+                    .uploadAttachments(message.attachments(),
+                            message.attachmentDimensions(),
+                            message.attachmentBlurHashes(),
+                            message.voiceNote());
             if (!additionalAttachments.isEmpty()) {
                 additionalAttachments.addAll(uploadedAttachments);
                 messageBuilder.withAttachments(additionalAttachments);